// src/nftmusic/dto/metadata.nft.dto.ts
export class MetadataNftDto {

  musicIpfsHash: string;

  imageIpfsHash: string;

  title: string; // 앨범타이틀

  name: string; // 노래제목

  artist: string; // 아티스트

  genres: string;  // 장르 파싱한 데이터

  description: string; // 설명

  playTime: number; // 재생시간

  minter: string;  // 최초발행자 지갑주소

  tokenId: string;
}